fix io

+++ cfdARCO/src/io_operators.cpp
@@ -1,4 +1,5 @@
 #include "io_operators.hpp"
+#include "cfdarcho_main.hpp"
 #include <chrono>
 #include <ctime>
 
@@ -8,43 +9,49 @@ using json = nlohmann::json;
 
 void store_history(const std::vector<Variable*>& vars_to_store, const Mesh2D* mesh, const fs::path& store_path) {
     std::time_t t = std::time(0);
-    std::tm* now = std::localtime(&t);
-    std::string timestamp = std::to_string(now->tm_year + 1900) + "_" +  std::to_string(now->tm_mon + 1) + "_" +
-            std::to_string(now->tm_mday) + "_" + std::to_string(now->tm_hour) + "_" + std::to_string(now->tm_min) +
-            "_" + std::to_string(now->tm_sec);
+    std::tm *now = std::localtime(&t);
+    std::string timestamp = std::to_string(now->tm_year + 1900) + "_" + std::to_string(now->tm_mon + 1) + "_" +
+                            std::to_string(now->tm_mday) + "_" + std::to_string(now->tm_hour) + "_" +
+                            std::to_string(now->tm_min) +
+                            "_" + std::to_string(now->tm_sec);
 
     auto store_dir = store_path / ("run_" + timestamp);
-    fs::create_directories(store_dir);
-
+    if (CFDArcoGlobalInit::get_rank() == 0) {
+        fs::create_directories(store_dir);
+    }
     for(auto var : vars_to_store) {
 
         auto store_var_dir = store_dir / var->name;
         fs::create_directories(store_var_dir);
 
         for (int i = 0; i < var->history.size() - 1; ++i) {
-            std::fstream file;
-            file.open(store_var_dir / (std::to_string(i) + ".bin"), std::ios_base::out | std::ios_base::binary);
-
-            if(!file.is_open())
-            {
-                std::cerr << "Unable to open the file" << std::endl;
-                return;
-            }
-
             auto grid_hist = to_grid(mesh, var->history[i]);
-            file.write(reinterpret_cast<char*>(grid_hist.data()), grid_hist.rows() * grid_hist.cols() * sizeof(double));
-            file.close();
+
+            if (CFDArcoGlobalInit::get_rank() == 0) {
+                std::fstream file;
+                file.open(store_var_dir / (std::to_string(i) + ".bin"), std::ios_base::out | std::ios_base::binary);
+
+                if (!file.is_open()) {
+                    std::cerr << "Unable to open the file" << std::endl;
+                    return;
+                }
+
+
+                file.write(reinterpret_cast<char *>(grid_hist.data()),
+                           grid_hist.rows() * grid_hist.cols() * sizeof(double));
+                file.close();
+            }
         }
     }
 
-    store_mesh(mesh, store_dir);
+    if (CFDArcoGlobalInit::get_rank() == 0) {
+        store_mesh(mesh, store_dir);
+        auto store_dir_latest = store_path / "run_latest";
+        fs::remove_all(store_dir_latest);
+        const auto copy_options = fs::copy_options::update_existing | fs::copy_options::recursive;
+        fs::copy(store_dir, store_dir_latest, copy_options);
 
-    auto store_dir_latest = store_path / "run_latest";
-    fs::remove_all(store_dir_latest);
-    const auto copy_options = fs::copy_options::update_existing | fs::copy_options::recursive;
-    fs::copy(store_dir, store_dir_latest, copy_options);
-
-
+    }
 }
 
 
@@ -69,7 +76,7 @@ void store_mesh(const Mesh2D* mesh, const fs::path& store_path) {
             {"nodes_id", nodes_id},
         });
     }
-    for (auto node : mesh->_nodes) {
+    for (auto node : mesh->_nodes_tot) {
         auto node_edges_arr = json::array();
         for (auto edge_id : node->_edges_id) {
             node_edges_arr.push_back(edge_id);
